OGD-228: Ensure MDC is cleared last
We want to ensure that the filter that clears the MDC is run last. This is so that previous filters can write log messages with the diagnostic context still set.

Thus, this filter must be matched early. Filters that run first run their post-request clean-up last, which is what we want here

Co-authored-by: Alex Newton <alex.newton@digital.cabinet-office.gov.uk>