EID-1944 Read test key as EC when supplied key not RSA
Private keys can be RSA or Elliptic Curve.
Try parsing test keys as EC as a fallback if the given key is not RSA.
This adds additional support for testing, which is required for EID-1944.

If the keys cannot be parsed as either, throw a InvalidKeySpecException for the RSA key.