HUB-912: Downgrade to Java 8
The MSA and the VSP need to stay compatible with Java 8 for the sake of
services.

These libraries are dependencies of saml-lib, which is a direct
dependency of the MSA and the VSP.

The versions of these libraries and saml-lib we publish need to remain
compiled to Java 8 byte code or the MSA and the VSP won't be able to use
them.